requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.2
django-environ==0.9.0
djangorestframework==3.14.0
OdooRPC==0.9.0
pytz==2022.5
sqlparse==0.4.3
